+++ vue.config.js
@@ -3,7 +3,7 @@ module.exports = {
 	publicPath: process.env.NODE_ENV === 'production' ? './' : '/',
 	lintOnSave: false,
 	devServer: {
-		open: process.env.VUE_APP_OPEN === 'false' ? false : true, // 自动打开浏览器
+		open: JSON.parse(process.env.VUE_APP_OPEN), // 自动打开浏览器
 		host: '0.0.0.0', // 真机模拟，使用
 		port: process.env.VUE_APP_PORT, // 前台代理端口号
 		https: false, // https： {type: Booleam}